PORTAL-119:  global search (#10)
* Deleted useless part

Deleted useless part

* drawed new search bar

drawed new search bar

* Added searchbox function

Added searchbox function

* Changed some style

Changed some style

* Adjusted inputbox position

Adjusted inputbox position

* drawed deleteIcon

drawed deleteIcon

* Added delete function to deleteIcon

Added delete function to deleteIcon

* Added delete icon behavior

Added delete icon behavior

* Added link to cardTitle

Added link to cardTitle
